hasher aangepast dat het instant een int wordt

+++ src/main/java/com/company/namingserver/Hasher.java
@@ -6,7 +6,7 @@ public class Hasher {
         long max = 2147483647;
         long min = -2147483647;
 
-        double result = ((input.hashCode()+max)*32768d)/(max+Math.abs(min));
+        long result = ((input.hashCode()+max)*32768)/(max+Math.abs(min));
 
         //double result = (IP.hashCode() + hostname.hashCode()+max)*(32768d/(max+abs(min)));
         //Eventueel hash maken over naam EN IP zodat je telkens een verschillende ID hebt
